refactor: reorganiza pacotes de exceções para uma melhor categorização
