+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/Problema.java
@@ -29,40 +29,6 @@ public class Problema {
         return new ProblemaBuilder();
     }
 
-    public int getStatus() {
-        return status;
-    }
-
-    // ----- Getters e Setters -----
-
-    public void setStatus(int status) {
-        this.status = status;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDetail() {
-        return detail;
-    }
-
-    public void setDetail(String detail) {
-        this.detail = detail;
-    }
-
     public static class ProblemaBuilder {
         private int status;
         private String type;
@@ -103,5 +69,39 @@ public class Problema {
         public Problema build() {
             return new Problema(status, type, title, detail, dataHora);
         }
+    }
+
+    // ----- Getters e Setters -----
+
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getDetail() {
+        return detail;
+    }
+
+    public void setDetail(String detail) {
+        this.detail = detail;
     }
 }
