+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/user/EmailAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.exceptions.auth.register;
+package br.com.gestorfinanceiro.exceptions.user;
 
 public class EmailAlreadyExistsException extends RuntimeException {
     public EmailAlreadyExistsException(String email) {
