+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/user/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.exceptions.auth.login;
+package br.com.gestorfinanceiro.exceptions.user;
 
 public class InvalidPasswordException extends RuntimeException {
     public InvalidPasswordException() {
