+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/user/InvalidUserIdException.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.exceptions;
+package br.com.gestorfinanceiro.exceptions.user;
 
 public class InvalidUserIdException extends RuntimeException {
     public InvalidUserIdException() {
