+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/user/UserOperationException.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.exceptions.auth;
+package br.com.gestorfinanceiro.exceptions.user;
 
 public class UserOperationException extends RuntimeException {
     public UserOperationException(String message, Throwable cause) {
