+++ backend/src/main/java/br/com/gestorfinanceiro/exceptions/user/UsernameAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.exceptions.auth.register;
+package br.com.gestorfinanceiro.exceptions.user;
 
 public class UsernameAlreadyExistsException extends RuntimeException {
     public UsernameAlreadyExistsException(String username) {
