+++ backend/src/test/java/br/com/gestorfinanceiro/controllers/AuthControllerTest/AuthControllerIntegrationTest.java
@@ -4,7 +4,7 @@ import br.com.gestorfinanceiro.TestDataUtil;
 import br.com.gestorfinanceiro.controller.AuthController;
 import br.com.gestorfinanceiro.dto.LoginDTO;
 import br.com.gestorfinanceiro.dto.UserDTO;
-import br.com.gestorfinanceiro.exceptions.auth.login.InvalidPasswordException;
+import br.com.gestorfinanceiro.exceptions.user.InvalidPasswordException;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.repositories.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
