+++ backend/src/test/java/br/com/gestorfinanceiro/services/AuthServiceTest/AuthServiceIntegrationTest.java
@@ -1,7 +1,7 @@
 package br.com.gestorfinanceiro.services.AuthServiceTest;
 
 import br.com.gestorfinanceiro.TestDataUtil;
-import br.com.gestorfinanceiro.exceptions.auth.register.EmailAlreadyExistsException;
+import br.com.gestorfinanceiro.exceptions.user.EmailAlreadyExistsException;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.models.enums.Roles;
 import br.com.gestorfinanceiro.repositories.UserRepository;
