+++ backend/src/test/java/br/com/gestorfinanceiro/services/DespesaServiceTest/DespesaServiceTest.java
@@ -2,11 +2,11 @@ package br.com.gestorfinanceiro.services.DespesaServiceTest;
 
 import br.com.gestorfinanceiro.dto.GraficoBarraDTO;
 import br.com.gestorfinanceiro.dto.GraficoPizzaDTO;
-import br.com.gestorfinanceiro.exceptions.InvalidDataException;
-import br.com.gestorfinanceiro.exceptions.InvalidUserIdException;
-import br.com.gestorfinanceiro.exceptions.InvalidUuidException;
 import br.com.gestorfinanceiro.exceptions.despesa.DespesaNotFoundException;
 import br.com.gestorfinanceiro.exceptions.despesa.DespesaOperationException;
+import br.com.gestorfinanceiro.exceptions.generalExceptions.InvalidDataException;
+import br.com.gestorfinanceiro.exceptions.generalExceptions.InvalidUuidException;
+import br.com.gestorfinanceiro.exceptions.user.InvalidUserIdException;
 import br.com.gestorfinanceiro.models.DespesaEntity;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.models.enums.DespesasCategorias;
