+++ backend/src/test/java/br/com/gestorfinanceiro/services/ReceitasServiceTest/ReceitaServiceIntegrationTest.java
@@ -2,10 +2,10 @@ package br.com.gestorfinanceiro.services.ReceitasServiceTest;
 
 import br.com.gestorfinanceiro.dto.GraficoBarraDTO;
 import br.com.gestorfinanceiro.dto.GraficoPizzaDTO;
-import br.com.gestorfinanceiro.exceptions.InvalidDataException;
-import br.com.gestorfinanceiro.exceptions.InvalidUserIdException;
-import br.com.gestorfinanceiro.exceptions.InvalidUuidException;
+import br.com.gestorfinanceiro.exceptions.generalExceptions.InvalidDataException;
+import br.com.gestorfinanceiro.exceptions.generalExceptions.InvalidUuidException;
 import br.com.gestorfinanceiro.exceptions.receita.ReceitaNotFoundException;
+import br.com.gestorfinanceiro.exceptions.user.InvalidUserIdException;
 import br.com.gestorfinanceiro.models.ReceitaEntity;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.models.enums.ReceitasCategorias;
